+++ src/view/Support/notice/style.css
@@ -7,3 +7,7 @@
     flex: 1;
     text-align: center;
 }
+
+#notice-wrapper .tr .td-title {
+    cursor: pointer;
+}
