+++ src/view/Support/qa/index.tsx
@@ -5,49 +5,8 @@ import { QaList } from '../../../types';
 import { usePagination } from '../../../hooks';
 import './style.css';
 import { useNavigate } from 'react-router-dom';
-import { SU_ABSOLUTE_QA_WRITE_PATH, SU_QA_DETAIL_PATH } from '../../../constants';
-
-// interface: 큐엔에이 리스트 아이템 //
-interface TableRowProps {
-    qas: QaList;
-};
-
-// component: QARow 컴포넌트 //
-// { qas }: TableRowProps
-function QaRow() {
-    // 데이터 넣은 후 요소 자리에 맞게 넣어주기
-
-    // function: 네비게이터 //
-    const navigator = useNavigate();
-
-    // event handler: 로우 클릭 시 디테일 페이지로 이동 //
-    //* 나중에 해당 번호 가져오기
-    const onDetailButtonHandler = () => {
-        navigator(SU_QA_DETAIL_PATH);
-    };
-
-    // render: QARow 컴포넌트 렌더링 //
-    return (
-        <>
-            <div className="tr" onClick={onDetailButtonHandler}>
-                <div className="td-no">10</div>
-                <div className="td-title">케이크가 주문대로 제작되지 않았습니다.</div>
-                <div className="td-writer">홍*동</div>
-                <div className="td-date">2024-10-04</div>
-                <div className="td-status">대기</div>
-            </div>
-            {/* 확인용 */}
-            <div className="tr">
-                <div className="td-no">10</div>
-                <div className="td-title">사장님이 주문 수락을 해주지 않아요.</div>
-                <div className="td-writer">남*현</div>
-                <div className="td-date">2024-10-04</div>
-                <div className="td-status">완료</div>
-            </div>
-        </>
-    )
-
-}
+import { SU_ABSOLUTE_QA_WRITE_PATH } from '../../../constants';
+import QaRow from '../../../components/qa_row';
 
 // component: support qa 컴포넌트 //
 export default function Qa() {
